fix: Add support for syncing .gizmo files in job bundle
This change adds support for syncing .gizmo files as part of the job bundle when the 'Include gizmos in job bundle' option is enabled. It includes finding all gizmo files used in the Nuke script, adding them to the job bundle, and setting up the NUKE_PATH environment variable to include the gizmo directory.

Issue: #144
Signed-off-by: Marin Maksutaj <19916629+MarinMaksutaj@users.noreply.github.com>